更新room

+++ ServerREST/src/room.h
@@ -53,7 +53,7 @@ public:
 
     int roomid;
 
-    int fee;
+    double fee;
 
     double currentTemp;
